+++ learning_assistant/api.py
@@ -4,6 +4,7 @@ Library for the learning_assistant app.
 import logging
 
 from django.conf import settings
+from django.contrib.auth import get_user_model
 from django.core.cache import cache
 from edx_django_utils.cache import get_cache_key
 from jinja2 import BaseLoader, Environment
@@ -11,7 +12,7 @@ from opaque_keys import InvalidKeyError
 
 from learning_assistant.constants import ACCEPTED_CATEGORY_TYPES, CATEGORY_TYPE_MAP
 from learning_assistant.data import LearningAssistantCourseEnabledData
-from learning_assistant.models import LearningAssistantCourseEnabled
+from learning_assistant.models import LearningAssistantCourseEnabled, LearningAssistantMessage
 from learning_assistant.platform_imports import (
     block_get_children,
     block_leaf_filter,
@@ -24,6 +25,7 @@ from learning_assistant.platform_imports import (
 from learning_assistant.text_utils import html_to_text
 
 log = logging.getLogger(__name__)
+User = get_user_model()
 
 
 def _extract_block_contents(child, category):
@@ -187,3 +189,38 @@ def get_course_id(course_run_id):
     course_data = get_cache_course_run_data(course_run_id, ['course'])
     course_key = course_data['course']
     return course_key
+
+
+def save_chat_message(courserun_key, user_id, chat_role, message):
+    """
+    Save the chat message to the database.
+    """
+    user = None
+    try:
+        user = User.objects.get(id=user_id)
+    except User.DoesNotExist as exc:
+        raise Exception("User does not exists.") from exc
+
+    # Save the user message to the database.
+    LearningAssistantMessage.objects.create(
+        course_id=courserun_key,
+        user=user,
+        role=chat_role,
+        content=message,
+
+    )
+
+
+def get_message_history(courserun_key, user, message_count):
+    """
+    Given a courserun key (CourseKey), user (User), and message count (int), return the associated message history.
+
+    Returns a number of messages equal to the message_count value.
+    """
+    # Explanation over the double reverse: This fetches the last message_count elements ordered by creating order DESC.
+    # Slicing the list in the model is an equivalent of adding LIMIT on the query.
+    # The result is the last chat messages for that user and course but in inversed order, so in order to flip them
+    # its first turn into a list and then reversed.
+    message_history = list(LearningAssistantMessage.objects.filter(
+        course_id=courserun_key, user=user).order_by('-created')[:message_count])[::-1]
+    return message_history