+++ learning_assistant/models.py
@@ -35,7 +35,30 @@ class LearningAssistantMessage(TimeStampedModel):
     .. pii_retirement: third_party
     """
 
+    USER_ROLE = 'user'
+    ASSISTANT_ROLE = 'assistant'
+
+    Roles = (
+        (USER_ROLE, USER_ROLE),
+        (ASSISTANT_ROLE, ASSISTANT_ROLE),
+    )
+
     course_id = CourseKeyField(max_length=255, db_index=True)
     user = models.ForeignKey(USER_MODEL, db_index=True, on_delete=models.CASCADE)
-    role = models.CharField(max_length=64)
+    role = models.CharField(choices=Roles, max_length=64)
     content = models.TextField()
+
+
+class LearningAssistantAuditTrial(TimeStampedModel):
+    """
+    This model stores the trial period for an audit learner using the learning assistant.
+
+    A LearningAssistantAuditTrial instance will be created on a per user basis,
+    when an audit learner first sends a message using Xpert LA.
+
+    .. no_pii: This model has no PII.
+    """
+
+    # Unique constraint since each user should only have one trial
+    user = models.ForeignKey(USER_MODEL, db_index=True, on_delete=models.CASCADE, unique=True)
+    start_date = models.DateTimeField()