+++ tests/test_views.py
@@ -1,10 +1,11 @@
 """
 Tests for the learning assistant views.
 """
+import datetime
 import json
 import sys
 from importlib import import_module
-from unittest.mock import MagicMock, patch
+from unittest.mock import MagicMock, call, patch
 
 import ddt
 from django.conf import settings
@@ -13,14 +14,18 @@ from django.http import HttpRequest
 from django.test import TestCase, override_settings
 from django.test.client import Client
 from django.urls import reverse
+from opaque_keys.edx.keys import CourseKey
+
+from learning_assistant.models import LearningAssistantMessage
 
 User = get_user_model()
 
 
-class TestClient(Client):
+class FakeClient(Client):
     """
     Allows for 'fake logins' of a user so we don't need to expose a 'login' HTTP endpoint.
     """
+
     def login_user(self, user):
         """
         Login as specified user.
@@ -63,14 +68,14 @@ class LoggedInTestCase(TestCase):
         Setup for tests.
         """
         super().setUp()
-        self.client = TestClient()
-        self.user = User(username='tester', email='tester@test.com')
+        self.client = FakeClient()
+        self.user = User(username='tester', email='tester@test.com', is_staff=True)
         self.user.save()
         self.client.login_user(self.user)
 
 
 @ddt.ddt
-class CourseChatViewTests(LoggedInTestCase):
+class TestCourseChatView(LoggedInTestCase):
     """
     Test for the CourseChatView
     """
@@ -82,6 +87,7 @@ class CourseChatViewTests(LoggedInTestCase):
     def setUp(self):
         super().setUp()
         self.course_id = 'course-v1:edx+test+23'
+        self.course_run_key = CourseKey.from_string(self.course_id)
 
         self.patcher = patch(
             'learning_assistant.api.get_cache_course_run_data',
@@ -150,15 +156,27 @@ class CourseChatViewTests(LoggedInTestCase):
         )
         self.assertEqual(response.status_code, 400)
 
+    @ddt.data(False, True)
     @patch('learning_assistant.views.render_prompt_template')
     @patch('learning_assistant.views.get_chat_response')
     @patch('learning_assistant.views.learning_assistant_enabled')
     @patch('learning_assistant.views.get_user_role')
     @patch('learning_assistant.views.CourseEnrollment.get_enrollment')
     @patch('learning_assistant.views.CourseMode')
+    @patch('learning_assistant.views.save_chat_message')
+    @patch('learning_assistant.views.chat_history_enabled')
     @override_settings(LEARNING_ASSISTANT_PROMPT_TEMPLATE='This is the default template')
     def test_chat_response_default(
-        self, mock_mode, mock_enrollment, mock_role, mock_waffle, mock_chat_response, mock_render
+        self,
+        enabled_flag,
+        mock_chat_history_enabled,
+        mock_save_chat_message,
+        mock_mode,
+        mock_enrollment,
+        mock_role,
+        mock_waffle,
+        mock_chat_response,
+        mock_render,
     ):
         mock_waffle.return_value = True
         mock_role.return_value = 'student'
@@ -168,9 +186,12 @@ class CourseChatViewTests(LoggedInTestCase):
         mock_render.return_value = 'Rendered template mock'
         test_unit_id = 'test-unit-id'
 
+        mock_chat_history_enabled.return_value = enabled_flag
+
         test_data = [
             {'role': 'user', 'content': 'What is 2+2?'},
-            {'role': 'assistant', 'content': 'It is 4'}
+            {'role': 'assistant', 'content': 'It is 4'},
+            {'role': 'user', 'content': 'And what else?'},
         ]
 
         response = self.client.post(
@@ -188,6 +209,14 @@ class CourseChatViewTests(LoggedInTestCase):
             'Rendered template mock',
             test_data,
         )
+
+        if enabled_flag:
+            mock_save_chat_message.assert_has_calls([
+                call(self.course_run_key, self.user.id, LearningAssistantMessage.USER_ROLE, test_data[-1]['content']),
+                call(self.course_run_key, self.user.id, LearningAssistantMessage.ASSISTANT_ROLE, 'Something else')
+            ])
+        else:
+            mock_save_chat_message.assert_not_called()
 
 
 @ddt.ddt
@@ -210,12 +239,12 @@ class LearningAssistantEnabledViewTests(LoggedInTestCase):
     )
     @ddt.unpack
     @patch('learning_assistant.views.learning_assistant_enabled')
-    def test_learning_assistant_enabled(self, mock_value, expected_value, mock_learning_assistant_enabled):
+    def test_learning_assistant_enabled(self, mock_value, message, mock_learning_assistant_enabled):
         mock_learning_assistant_enabled.return_value = mock_value
         response = self.client.get(reverse('enabled', kwargs={'course_run_id': self.course_id}))
 
         self.assertEqual(response.status_code, 200)
-        self.assertEqual(response.data, {'enabled': expected_value})
+        self.assertEqual(response.data, {'enabled': message})
 
     @patch('learning_assistant.views.learning_assistant_enabled')
     def test_invalid_course_id(self, mock_learning_assistant_enabled):
@@ -223,3 +252,129 @@ class LearningAssistantEnabledViewTests(LoggedInTestCase):
         response = self.client.get(reverse('enabled', kwargs={'course_run_id': self.course_id+'+invalid'}))
 
         self.assertEqual(response.status_code, 400)
+
+
+@ddt.ddt
+class LearningAssistantMessageHistoryViewTests(LoggedInTestCase):
+    """
+    Tests for the LearningAssistantMessageHistoryView
+    """
+
+    def setUp(self):
+        super().setUp()
+        self.course_id = 'course-v1:edx+test+23'
+
+    @patch('learning_assistant.views.learning_assistant_enabled')
+    def test_invalid_course_id(self, mock_learning_assistant_enabled):
+        mock_learning_assistant_enabled.return_value = True
+        response = self.client.get(reverse('enabled', kwargs={'course_run_id': self.course_id+'+invalid'}))
+
+        self.assertEqual(response.status_code, 400)
+
+    @patch('learning_assistant.views.learning_assistant_enabled')
+    def test_course_waffle_inactive(self, mock_waffle):
+        mock_waffle.return_value = False
+        message_count = 5
+        response = self.client.get(
+            reverse('message-history', kwargs={'course_run_id': self.course_id})+f'?message_count={message_count}',
+            content_type='application/json'
+        )
+        self.assertEqual(response.status_code, 403)
+
+    @patch('learning_assistant.views.learning_assistant_enabled')
+    def test_learning_assistant_not_enabled(self, mock_learning_assistant_enabled):
+        mock_learning_assistant_enabled.return_value = False
+        message_count = 5
+        response = self.client.get(
+            reverse('message-history', kwargs={'course_run_id': self.course_id})+f'?message_count={message_count}',
+            content_type='application/json'
+        )
+
+        self.assertEqual(response.status_code, 403)
+
+    @patch('learning_assistant.views.learning_assistant_enabled')
+    @patch('learning_assistant.views.get_user_role')
+    @patch('learning_assistant.views.CourseEnrollment.get_enrollment')
+    @patch('learning_assistant.views.CourseMode')
+    def test_user_no_enrollment_not_staff(self, mock_mode, mock_enrollment, mock_role, mock_waffle):
+        mock_waffle.return_value = True
+        mock_role.return_value = 'student'
+        mock_mode.VERIFIED_MODES = ['verified']
+        mock_enrollment.return_value = None
+
+        message_count = 5
+        response = self.client.get(
+            reverse('message-history', kwargs={'course_run_id': self.course_id})+f'?message_count={message_count}',
+            content_type='application/json'
+        )
+        self.assertEqual(response.status_code, 403)
+
+    @patch('learning_assistant.views.learning_assistant_enabled')
+    @patch('learning_assistant.views.get_user_role')
+    @patch('learning_assistant.views.CourseEnrollment.get_enrollment')
+    @patch('learning_assistant.views.CourseMode')
+    def test_user_audit_enrollment_not_staff(self, mock_mode, mock_enrollment, mock_role, mock_waffle):
+        mock_waffle.return_value = True
+        mock_role.return_value = 'student'
+        mock_mode.VERIFIED_MODES = ['verified']
+        mock_enrollment.return_value = MagicMock(mode='audit')
+
+        message_count = 5
+        response = self.client.get(
+            reverse('message-history', kwargs={'course_run_id': self.course_id})+f'?message_count={message_count}',
+            content_type='application/json'
+        )
+        self.assertEqual(response.status_code, 403)
+
+    @patch('learning_assistant.views.learning_assistant_enabled')
+    @patch('learning_assistant.views.get_user_role')
+    @patch('learning_assistant.views.CourseEnrollment.get_enrollment')
+    @patch('learning_assistant.views.CourseMode')
+    @patch('learning_assistant.views.get_course_id')
+    def test_learning_message_history_view_get(
+        self,
+        mock_get_course_id,
+        mock_mode,
+        mock_enrollment,
+        mock_role,
+        mock_waffle
+    ):
+        mock_waffle.return_value = True
+        mock_role.return_value = 'student'
+        mock_mode.VERIFIED_MODES = ['verified']
+        mock_enrollment.return_value = MagicMock(mode='verified')
+
+        LearningAssistantMessage.objects.create(
+            course_id=self.course_id,
+            user=self.user,
+            role='staff',
+            content='Older message',
+            created=datetime.date(2024, 10, 1)
+        )
+
+        LearningAssistantMessage.objects.create(
+            course_id=self.course_id,
+            user=self.user,
+            role='staff',
+            content='Newer message',
+            created=datetime.date(2024, 10, 3)
+        )
+
+        db_messages = LearningAssistantMessage.objects.all().order_by('created')
+        db_messages_count = len(db_messages)
+
+        mock_get_course_id.return_value = self.course_id
+        response = self.client.get(
+            reverse('message-history', kwargs={'course_run_id': self.course_id})+f'?message_count={db_messages_count}',
+            content_type='application/json'
+        )
+        data = response.data
+
+        # Ensure same number of entries
+        self.assertEqual(len(data), db_messages_count)
+
+        # Ensure values are as expected
+        for i, message in enumerate(data):
+            self.assertEqual(message['role'], db_messages[i].role)
+            self.assertEqual(message['content'], db_messages[i].content)
+            self.assertEqual(message['timestamp'], db_messages[i].created.isoformat())